Error handling in OWIN middleware
Allows handling of exceptions and generation of problem details responses directly in the OWIN middleware (see #1).

`ProblemDetailsMiddlewareOptions` has a new `ExceptionHandler` property, which is a delegate that is invoked when an exception bubbles up from the OWIN pipeline. The delegate can choose to generate a problem details object that will be returned. If no problem details object is generated, the exception is rethrown.