Allow Web API exceptions to be handled by OWIN pipeline
These changes allow the default catch-all `IExceptionHandler` service in the Web API configuration to be replaced with one that will cause all unhandled exceptions to bubble up from the Web API pipeline into the host app's request pipeline.

This allows all exceptions to be handled in the OWIN pipeline instead of in the Web API pipeline.