Handle model state errors occurring after action filter
If the model state for a request becomes invalid after ProblemDetailsActionFilterAttribute.OnActionExecuting has run (e.g. if another action filter or a controller adds an error to the model state), create a validation problem details response in ProblemDetailsActionFilterAttribute.OnActionExecuted. This is only done if the response code is 400, to ensure that e.g. 500 errors are not obscured if the model state is also invalid.